Use BindGlobal for internal helpers
There is no need in general to use DeclareGlobalFunction at all,
but it can be handy to make it clear what the 'public' functions
of a package are, i.e., those functions users are supposed to be
calling. AutoDoc also makes it easy to attach documentation to it,
so that's handy. But using it for everything is overkill. Most
internal functions can just be set via BindGlobal (strictly
speaking even that is not necessary, and one could just use
regular assignment, that is, `helper:=function() ... end` -- but
using BindGlobal makes those variables read-only, which helps
avoid all kinds of bugs caused by accidentally overwriting such
helpers with other code.