Use A[i,j] instead of A[i][j] and other MatrixObj tweaks
Note: one should not blindly replace all accesses of the form
`foo[a][b]` by `foo[a,b]`, but rather limit this to the case
where `foo` really is a matrix (and not a list of matrices, or
a list of vector, etc.).

In addition, use OneOfBaseDomain / ZeroOfBaseDomain, resp. just
directly ask the base file for a one / zero element.